Fix eslint errors and move files to src folder
